requirements.txt
streamlit
srt
pydub
python-dotenv
openai
imageio[ffmpeg]